tests/requirements.txt
pytest==7.4.3
pyyaml==6.0.1

